pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autodocgen"
version = "0.1.0"
description = "AI-powered CLI tool that generates documentation from code comments"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "AutoChamp"}]
keywords = ["documentation", "ai", "cli", "generator"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "openai>=1.0.0",
    "jinja2>=3.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "toml>=0.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
autodocgen = "autodocgen.cli:main"

[tool.setuptools]
package-dir = {"" = "."}
include-package-data = true

[tool.setuptools.packages]
find = {}

[tool.setuptools.package-data]
autodocgen = ["templates/*.j2"]